% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/action.R
\name{action}
\alias{action}
\title{Build and execute a `unifir_script`}
\usage{
action(script, write = TRUE, exec = TRUE, quit = TRUE)
}
\arguments{
\item{script}{The `unifir_script` object to build and execute.}

\item{write}{Boolean: Write the generated script to a file?}

\item{exec}{Boolean: Execute the script inside of the Unity project? Note
that if `write = FALSE`, `exec` cannot be `TRUE`.}

\item{quit}{Boolean: Quit Unity after execution?}
}
\description{
Build and execute a `unifir_script`
}
