% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_prop}
\alias{add_prop}
\title{Add a prop to a unifir script}
\usage{
add_prop(script, prop, exec = TRUE)
}
\arguments{
\item{script}{A script object (from [make_script]) to append the prop to.}

\item{prop}{A [unifir_prop] object to add to the script}

\item{exec}{Logical: Should the method created by the prop be called in the
MainFunc method?}
}
\description{
Add a prop to a unifir script
}
