% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_texture.R
\name{add_texture}
\alias{add_texture}
\title{Add a Texture2D layer to a terrain tile object}
\usage{
add_texture(script, method_name = NULL, exec = FALSE)
}
\arguments{
\item{script}{A `unifir_script` object, created by [make_script] or returned
by an `add_prop_*` function.}

\item{method_name}{The internal name to use for the C# method created. Will
be randomly generated if not set.}

\item{exec}{Logical: Should the C# method be included in the set executed by
MainFunc?}
}
\description{
Add a Texture2D layer to a terrain tile object
}
