% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{available_assets}
\alias{available_assets}
\title{Vector of assets unifir can download and import}
\format{
A character vector with 13 elements, each representing an asset
which can be imported.
}
\source{
\url{https://github.com/mikemahoney218/unity_assets}
}
\usage{
available_assets
}
\description{
This object contains the set of assets unifir is able to download and import
(through \link{get_asset} and \link{import_asset}). These objects are all released
under permissive open-source licenses (currently, either CC-0 1.0 or MIT).
More information on the assets may be found at
https://github.com/mikemahoney218/unity_assets .
}
\keyword{datasets}
