% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_if_not}
\alias{create_if_not}
\title{Create directory if it doesn't exist}
\usage{
create_if_not(path, recur = FALSE)
}
\arguments{
\item{path}{The path to be created}

\item{recur}{Boolean: create directories recursively?}
}
\description{
Create directory if it doesn't exist
}
