% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_terrain.R
\name{create_terrain}
\alias{create_terrain}
\title{Create a terrain tile with optional image overlay}
\usage{
create_terrain(
  script,
  method_name = NULL,
  heightmap_path,
  x_pos,
  z_pos,
  width,
  height,
  length,
  heightmap_resolution,
  texture_path = "",
  exec = TRUE
)
}
\arguments{
\item{script}{A `unifir_script` object, created by [make_script] or returned
by an `add_prop_*` function.}

\item{method_name}{The internal name to use for the C# method created. Will
be randomly generated if not set.}

\item{heightmap_path}{The file path to the heightmap to import as terrain.}

\item{x_pos, z_pos}{The position of the corner of the terrain.}

\item{width, height, length}{The dimensions of the terrain tile, in linear units.}

\item{heightmap_resolution}{The resolution of the heightmap image.}

\item{texture_path}{Optional: the file path to the image to use as a terrain
overlay.}

\item{exec}{Logical: Should the C# method be included in the set executed by
MainFunc?}
}
\description{
Create a terrain tile with optional image overlay
}
