% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{create_unity_project}
\alias{create_unity_project}
\title{Create a new Unity project.}
\usage{
create_unity_project(path, unity = NULL)
}
\arguments{
\item{path}{The path to create a new Unity project at.}

\item{unity}{The path to the Unity executable on your system (importantly,
_not_ the UnityHub executable). If `NULL`, checks to see if the environment
variable or option `unifir_unity_path` is set; if so, uses that path
(preferring the environment variable over the option if the two disagree).}
}
\value{
TRUE, invisibly.
}
\description{
Create a new Unity project.
}
