% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_unity.R
\name{find_unity}
\alias{find_unity}
\title{Find the Unity executable on a machine.}
\usage{
find_unity(unity = NULL)
}
\arguments{
\item{unity}{Character: If provided, this function will quote the provided
string (if necessary) and return it without searching other paths (or
confirming that the provided path exists).}
}
\description{
If the path to Unity is not provided to a function, this function is invoked
to attempt to find it. To do so, it goes through the following steps:
}
\details{
1. Attempt to load the "unifir_unity_path" environment variable.
2. Attempt to load the "unifir_unity_path" option.
}
