% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_asset.R
\name{import_asset}
\alias{import_asset}
\title{Import assets into Unity.}
\usage{
import_asset(script, package, exec = TRUE, lazy = TRUE)
}
\arguments{
\item{script}{A `unifir_script` object, created by [make_script] or returned
by an `add_prop_*` function.}

\item{package}{The file path to the asset to import. If a directory, the
entire directory will be recursively copied.}

\item{exec}{Logical: Should the C# method be included in the set executed by
MainFunc?}

\item{lazy}{Boolean: if TRUE, unifir will attempt to only copy the files
once; if FALSE, unifir will copy the files as many times as requested,
overwriting pre-existing files each time.}
}
\value{
`script` with a new prop.
}
\description{
Import assets into Unity.
}
