% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_png.R
\name{load_png}
\alias{load_png}
\title{Create a Texture2D from a PNG file}
\usage{
load_png(script, method_name = NULL)
}
\arguments{
\item{script}{A `unifir_script` object, created by [make_script] or returned
by an `add_prop_*` function.}

\item{method_name}{The internal name to use for the C# method created. Will
be randomly generated if not set.}
}
\description{
Create a Texture2D from a PNG file
}
