% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scene_management.R
\name{load_scene}
\alias{load_scene}
\title{Load a scene in a Unity project.}
\usage{
load_scene(script, scene_name, method_name = NULL)
}
\arguments{
\item{script}{A `unifir_script` object, created by [make_script] or returned
by an `add_prop_*` function.}

\item{scene_name}{The name of the scene to load.}

\item{method_name}{The internal name to use for the C# method created. Will
be randomly generated if not set.}
}
\description{
Load a scene in a Unity project.
}
