% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scene_management.R
\name{save_scene}
\alias{save_scene}
\title{Save a scene in a Unity project.}
\usage{
save_scene(script, scene_name = NULL, method_name = NULL, exec = TRUE)
}
\arguments{
\item{script}{A `unifir_script` object, created by [make_script] or returned
by an `add_prop_*` function.}

\item{scene_name}{The name to save the scene to.}

\item{method_name}{The internal name to use for the C# method created. Will
be randomly generated if not set.}

\item{exec}{Logical: Should the C# method be included in the set executed by
MainFunc?}
}
\description{
Save a scene in a Unity project.
}
