% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scene_management.R
\name{set_active_scene}
\alias{set_active_scene}
\title{Set a single scene to active.}
\usage{
set_active_scene(script, scene_name = NULL)
}
\arguments{
\item{script}{A `unifir_script` object, created by [make_script] or returned
by an `add_prop_*` function.}

\item{scene_name}{The name of the scene to set as the active scene.}
}
\description{
Set a single scene to active.
}
