% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/action.R
\name{set_script_defaults}
\alias{set_script_defaults}
\title{Fill in plot holes in a script}
\usage{
set_script_defaults(script, debug)
}
\arguments{
\item{script}{The unifir_script to fill elements of}

\item{debug}{Boolean: run in debug mode?}
}
\description{
Fill in plot holes in a script
}
