% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unifir-package.R
\docType{package}
\name{unifir-package}
\alias{unifir}
\alias{unifir-package}
\title{unifir: A Unifying API for Calling the 'Unity' '3D' Video Game Engine}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Functions for the creation and manipulation of scenes and objects within the 'Unity' '3D' video game engine (<https://unity.com/>). Specific focuses include the creation and import of terrain data and 'GameObjects' as well as scene management.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ropensci.github.io/unifir/}
  \item \url{https://github.com/ropensci/unifir}
  \item Report bugs at \url{https://github.com/ropensci/unifir/issues}
}

}
\author{
\strong{Maintainer}: Michael Mahoney \email{mike.mahoney.218@gmail.com} (\href{https://orcid.org/0000-0003-2402-304X}{ORCID})

Other contributors:
\itemize{
  \item Will Jones (Will reviewed the package (v. 0.2.0) for rOpenSci, see <https://github.com/ropensci/software-review/issues/521>) [reviewer]
  \item Tan Tran (Tan reviewed the package (v. 0.2.0) for rOpenSci, see <https://github.com/ropensci/software-review/issues/521>) [reviewer]
}

}
\keyword{internal}
