% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_path.R
\name{validate_path}
\alias{validate_path}
\title{Validate a file path exists}
\usage{
validate_path(script, path, method_name = NULL)
}
\arguments{
\item{script}{A `unifir_script` object, created by [make_script] or returned
by an `add_prop_*` function.}

\item{path}{The file path to validate}

\item{method_name}{The internal name to use for the C# method created. Will
be randomly generated if not set.}
}
\description{
Validate a file path exists
}
