% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{waiver}
\alias{waiver}
\title{A waiver object.}
\usage{
waiver()
}
\description{
This function is borrowed from ggplot2. It creates a "flag" object indicating
that a value has been intentionally left blank (because it will be filled in
by something else). Often, a function argument being missing or `NULL` will
result in an error, while passing `waiver()` will cause the function to look
elsewhere in the script for an acceptable value.
}
\examples{
waiver()

}
\references{
H. Wickham. ggplot2: Elegant Graphics for Data Analysis. Springer-Verlag New York, 2016.
}
