% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_players.R
\name{get_players}
\alias{get_players}
\title{Download player prefabs for Unity}
\usage{
get_players(directory = NULL)
}
\arguments{
\item{directory}{Optionally, the directory to extract the downloaded models
in. If NULL, the default, saves to `tools::R_user_dir("unifir")`.}
}
\description{
This is a simple helper function downloading the player prefabs stored at
https://github.com/boaheck/TheFirstPerson .
}
