% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_asset.R
\name{import_asset}
\alias{import_asset}
\title{Import assets into Unity.}
\usage{
import_asset(script, asset_path, lazy = TRUE)
}
\arguments{
\item{script}{A `unifir_script` object, created by [make_script] or returned
by an `add_prop_*` function.}

\item{asset_path}{The file path to the asset to import. If a directory, the
entire directory will be recursively copied. Note that this function doesn't
have a `method_name` argument: the `asset_path` is used as the method name.}

\item{lazy}{Boolean: if TRUE, unifir will attempt to only copy the files
once per run of a script; if FALSE, unifir will copy the files as many times as
requested, overwriting pre-existing files each time.}
}
\value{
`script` with a new prop.
}
\description{
Import assets into Unity.
}
\seealso{
Other props: 
\code{\link{add_default_player}()},
\code{\link{add_light}()},
\code{\link{add_prop}()},
\code{\link{add_texture}()},
\code{\link{create_terrain}()},
\code{\link{instantiate_prefab}()},
\code{\link{load_png}()},
\code{\link{load_scene}()},
\code{\link{new_scene}()},
\code{\link{read_raw}()},
\code{\link{save_scene}()},
\code{\link{set_active_scene}()},
\code{\link{validate_path}()}
}
\concept{props}
