% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scene_management.R
\name{new_scene}
\alias{new_scene}
\title{Create a new scene in a Unity project.}
\usage{
new_scene(
  script,
  setup = c("EmptyScene", "DefaultGameObjects"),
  mode = c("Additive", "Single"),
  method_name = NULL
)
}
\arguments{
\item{script}{A `unifir_script` object, created by [make_script] or returned
by an `add_prop_*` function.}

\item{setup}{One of "EmptyScene"
("No game objects are added to the new Scene.") or "DefaultGameObjects"
("Adds default game objects to the new Scene (a light and camera).")}

\item{mode}{One of "Additive"
("The newly created Scene is added to the current open Scenes.") or "Single"
("All current open Scenes are closed and the newly created Scene are
opened.")}

\item{method_name}{The internal name to use for the C# method created. Will
be randomly generated if not set.}
}
\description{
Create a new scene in a Unity project.
}
