% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unifir-package.R
\docType{package}
\name{unifir-package}
\alias{unifir}
\alias{unifir-package}
\title{unifir: A Unifying API for Calling 'Unity' from R}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Functions for the creation and manipulation of scenes and objects within the 'Unity' '3D' video game engine (<https://unity.com/>). Specific focuses include the creation and import of terrain data and 'GameObjects' as well as scene management.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mikemahoney218.github.io/unifir/}
  \item \url{https://github.com/mikemahoney218/unifir}
  \item Report bugs at \url{https://github.com/mikemahoney218/unifir/issues}
}

}
\author{
\strong{Maintainer}: Michael Mahoney \email{mike.mahoney.218@gmail.com} (\href{https://orcid.org/0000-0003-2402-304X}{ORCID})

}
\keyword{internal}
