% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop.R
\name{unifir_prop}
\alias{unifir_prop}
\title{The class for unifir prop objects}
\usage{
unifir_prop(prop_file, method_name, method_type, parameters, build, using)
}
\arguments{
\item{prop_file}{The system location for the C# template file}

\item{method_name}{The name of the method, in C# code}

\item{method_type}{The type of the method (usually matches its file name);
scripts can have multiple versions of the same method, each with different
method_name values, all sharing the same method_type.}

\item{parameters}{Method-specific parameters, typically used in the build
stage.}

\item{build}{A function that takes two arguments, `script` and `prop`, and
uses those to construct the C# method.}

\item{using}{A character vector of imports required for the method.}
}
\value{
An R6 object of class `unifir_prop`
}
\description{
The class for unifir prop objects
}
