% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_path.R
\name{validate_path}
\alias{validate_path}
\alias{validate_single_path}
\title{Validate a file path exists}
\usage{
validate_path(script, path, method_name = NULL, exec = FALSE)

validate_single_path(script, path, method_name = NULL, exec = TRUE)
}
\arguments{
\item{script}{A `unifir_script` object, created by [make_script] or returned
by an `add_prop_*` function.}

\item{path}{The file path to validate}

\item{method_name}{The internal name to use for the C# method created. Will
be randomly generated if not set.}

\item{exec}{Logical: Should the C# method be included in the set executed by
MainFunc?}
}
\description{
`validate_path` creates a generic C# method which takes a single argument
and checks to make sure it exists. Your C# code calling the method must
provide the path to validate.
`validate_single_path` hard-codes the path to check in the C# code. This
allows you to specify the path to check from R.
}
