% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{waiver}
\alias{waiver}
\title{A waiver object.}
\usage{
waiver()
}
\value{
An empty list of class \code{waiver}.
}
\description{
This function is borrowed from ggplot2. It creates a "flag" object indicating
that a value has been intentionally left blank (because it will be filled in
by something else). Often, a function argument being missing or \code{NULL} will
result in an error, while passing \code{waiver()} will cause the function to look
elsewhere in the script for an acceptable value.
}
\examples{
waiver()

}
\references{
H. Wickham. ggplot2: Elegant Graphics for Data Analysis. Springer-Verlag New York, 2016.
}
\seealso{
Other utilities: 
\code{\link{add_default_player}()},
\code{\link{add_prop}()},
\code{\link{create_unity_project}()},
\code{\link{find_unity}()},
\code{\link{get_asset}()},
\code{\link{load_png}()},
\code{\link{load_scene}()},
\code{\link{new_scene}()},
\code{\link{read_raw}()},
\code{\link{save_scene}()},
\code{\link{set_active_scene}()},
\code{\link{validate_path}()}
}
\concept{utilities}
